Upload a model, a tokenizer algorithm
Need to probably get a different tokenizer algorithm and make a tokenizer. The model should work good for the purpose of the demo.
I am uploading so that anyone can try to work with it, I cant get the emulator working. Technically, this shouldnt work because the library wont be able to find the model (gpt-2). If you want to see the documentation for it its on hugging face distilgpt2. There is a tokenizer and vocab file.