OK. I got it to actually do something
It took many many many hours, but I got the plugin to work with the AI. I significantly improved my knowledge. The AI can actually generate something, its nonsense, but with tuning and changing some things we can get it to autocomplete. Its not the final AI, but it does pretty good to show its feasible to run a language model locally on a mobile device.